Move EmptyStateTemplate to top level components dir
