+++ ui/apps/platform/src/Components/EmptyStateTemplate/EmptyStateTemplate.tsx
