+++ ui/apps/platform/src/Components/EmptyStateTemplate/index.ts
