+++ ui/apps/platform/src/Components/PatternFly/WidgetErrorEmptyState.tsx
@@ -3,7 +3,7 @@ import { Flex } from '@patternfly/react-core';
 import { ExclamationCircleIcon } from '@patternfly/react-icons';
 import { SVGIconProps } from '@patternfly/react-icons/dist/js/createIcon';
 
-import EmptyStateTemplate from 'Components/PatternFly/EmptyStateTemplate';
+import EmptyStateTemplate from 'Components/EmptyStateTemplate';
 
 export type WidgetErrorEmptyStateProps = {
     children: ReactNode;
