+++ ui/apps/platform/src/Components/TableStateTemplates/TbodyError.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { ExclamationCircleIcon } from '@patternfly/react-icons';
 
-import EmptyStateTemplate from 'Components/PatternFly/EmptyStateTemplate';
-import { EmptyStateTemplateProps } from 'Components/PatternFly/EmptyStateTemplate/EmptyStateTemplate';
+import EmptyStateTemplate from 'Components/EmptyStateTemplate';
+import { EmptyStateTemplateProps } from 'Components/EmptyStateTemplate/EmptyStateTemplate';
 import { getAxiosErrorMessage } from 'utils/responseErrorUtils';
 import { TbodyFullCentered } from './TbodyFullCentered';
