+++ ui/apps/platform/src/Components/TableStateTemplates/TbodyFilteredEmpty.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { Button, ButtonProps, Text } from '@patternfly/react-core';
 import { SearchIcon } from '@patternfly/react-icons';
 
-import EmptyStateTemplate from 'Components/PatternFly/EmptyStateTemplate';
-import { EmptyStateTemplateProps } from 'Components/PatternFly/EmptyStateTemplate/EmptyStateTemplate';
+import EmptyStateTemplate from 'Components/EmptyStateTemplate';
+import { EmptyStateTemplateProps } from 'Components/EmptyStateTemplate/EmptyStateTemplate';
 
 import { TbodyFullCentered } from './TbodyFullCentered';
