+++ ui/apps/platform/src/Containers/Collections/UnsupportedCollectionState.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { ExpandableSection, List, ListItem } from '@patternfly/react-core';
 import { InfoCircleIcon } from '@patternfly/react-icons';
 
-import EmptyStateTemplate from 'Components/PatternFly/EmptyStateTemplate';
+import EmptyStateTemplate from 'Components/EmptyStateTemplate';
 import { CollectionParseError } from './converter';
 
 export type UnsupportedCollectionStateProps = {
