+++ ui/apps/platform/src/Containers/NetworkGraph/components/EmptyUnscopedState.tsx
@@ -3,7 +3,7 @@ import { Bullseye, Icon, Text } from '@patternfly/react-core';
 import { ModuleIcon } from '@patternfly/react-icons';
 import { SVGIconProps } from '@patternfly/react-icons/dist/esm/createIcon';
 
-import EmptyStateTemplate from 'Components/PatternFly/EmptyStateTemplate';
+import EmptyStateTemplate from 'Components/EmptyStateTemplate';
 
 function EmptyIcon(props: SVGIconProps) {
     return (
