+++ ui/apps/platform/src/Containers/VulnMgmt/RiskAcceptance/DeferredCVEs/DeferredCVEsTable.tsx
@@ -22,7 +22,7 @@ import { UsePaginationResult } from 'hooks/patternfly/usePagination';
 import usePermissions from 'hooks/usePermissions';
 import useAuthStatus from 'hooks/useAuthStatus';
 import { SearchFilter } from 'types/search';
-import EmptyStateTemplate from 'Components/PatternFly/EmptyStateTemplate';
+import EmptyStateTemplate from 'Components/EmptyStateTemplate';
 import { GetSortParams } from 'hooks/patternfly/useTableSort';
 import { Vulnerability, EmbeddedImageScanComponent } from '../imageVulnerabilities.graphql';
 import { DeferredCVEsToBeAssessed } from './types';
